更に修正しフルスクリーン化に対応

+++ Action/AttackPlayerOnce.cpp
@@ -14,7 +14,7 @@ AttackPlayerOnce::~AttackPlayerOnce()
 {
 }
 
-void AttackPlayerOnce::Attack(const Vector3& _ownerPosition,const int& _direction)
+void AttackPlayerOnce::Attack(const Vector3& _ownerPosition,const float& _direction)
 {
 	new WeaponRotationAnimationPlayer(_ownerPosition, _direction,0,120,0);
 }
