+++ Action/AttackPlayerThird.cpp
@@ -13,7 +13,7 @@ AttackPlayerThird::~AttackPlayerThird()
 {
 }
 
-void AttackPlayerThird::Attack(const Vector3 & _ownerPosition, const int& _direction)
+void AttackPlayerThird::Attack(const Vector3 & _ownerPosition, const float& _direction)
 {
 	new WeaponRotationAnimationPlayer(_ownerPosition, _direction, 0, 120, 2);
