+++ Action/WeaponRotationAnimationPlayer.cpp
@@ -4,7 +4,7 @@
 #include "ColliderComponent.h"
 #include "RotateComponent.h"
 
-WeaponRotationAnimationPlayer::WeaponRotationAnimationPlayer(const Vector3& _pos, const int& _direction, const int& _waitCount, const int& _lifeCount, const int& _moveDistanceStage) :
+WeaponRotationAnimationPlayer::WeaponRotationAnimationPlayer(const Vector3& _pos, const float& _direction, const int& _waitCount, const int& _lifeCount, const int& _moveDistanceStage) :
 	WeaponBase(_pos, _waitCount, _lifeCount, _direction,
 		Tag::PlayerWeaponTag,3, Vector3(75, 75, 75),0.02f, "Assets/Model/Weapon/SK_Blunt_SpikedClub.gpmesh", "Assets/Model/Weapon/SK_Blunt_SpikedClub.gpskel"),
 	rotateSpeed(50),
